requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.4
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.0
h5py==3.11.0
html5lib==1.1
huggingface-hub==0.24.6
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
multitasking==0.0.11
namex==0.0.8
networkx==3.3
newsapi-python==0.2.7
nltk==3.9.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
protobuf==4.25.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
yfinance==0.2.42
